test(4theta): Add 4Theta to autoregressive test. Move Enums to top init file
