+++ darts/models/__init__.py
@@ -15,22 +15,3 @@ from .fft import FFT
 
 # Regression
 from .standard_regression_model import StandardRegressionModel
-
-# Enums
-from enum import Enum
-
-
-class Season(Enum):
-    MULTIPLICATIVE = 'multiplicative'
-    ADDITIVE = 'additive'
-    NONE = None
-
-
-class Trend(Enum):
-    LINEAR = 'linear'
-    EXPONENTIAL = 'exponential'
-
-
-class Model(Enum):
-    MULTIPLICATIVE = 'multiplicative'
-    ADDITIVE = 'additive'
